+++ CHANGELOG.md
@@ -4,3 +4,4 @@ here's only the changelog for the version in development
 # v2.1
 
 * proofed against loading empty strings
+* Python 3.6 is supported