+++ CHANGELOG.md
@@ -5,3 +5,4 @@ here's only the changelog for the version in development
 
 * proofed against loading empty strings
 * Python 3.6 is supported
+* minor speed improvements