+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog is kept at [GitHub](https://github.com/Dronehub/minijson/releases),
 here's only the changelog for the version in development
 
-# v2.2
+# v2.3
 
-* added support for PyPy and Python 3.5
+* loads will now take any object that can provide 
+  it's `__bytes__`